Block Diagram
READER SECTION (VIRTUAL TERMINAL):
This section contains a virtual terminal. It is used as a reader where input is typed by the user in the virtual terminal using text. It reads input data serially. This virtual terminal is working at baud rate of 9600 bps. Practically it can be replaced by RFID that uses electromagnet to transfer 12 digit character code data from tag to reader. In our program we used 15 digit Register Number given by college as input in virtual terminal.
CONTROL SECTION (8051 MICROCONTROLLER):
8051 microcontroller is used for controlling the complete process of this project. Here by using 8051 we are receiving virtual terminal data and sending status or messages to LCD.
DISPLAY SECTION (16x2 LCD):
A 16x2 LCD is used in this project for displaying messages on it.